+++ spec/system/google_login_spec.rb
@@ -1,27 +1,15 @@
 require "rails_helper"
 
-RSpec.describe "Googleログイン", type: :request do
-  it "ユーザーを作成し、ログインする" do
-    expect do
-      get "/auth/google_oauth2/callback"
-    end.to change(User, :count).by(1)
-
-    expect(session[:user_id]).to be_present
+RSpec.describe "Googleログイン", type: :system do
+  before do
+    driven_by(:rack_test)
   end
 
-  context "既存ユーザーがいる場合 " do
-    let!(:user) do
-      User.create!(
-        email_address: "test@example.com",
-        provider: "google_oauth2",
-        uid: "123456789"
-      )
-    end
+  scenario "ログインすると、ふりかえりの入力画面に遷移する" do
+    visit root_path
 
-    it "新しく追加しない" do
-      expect do
-        get "/auth/google_oauth2/callback"
-      end.not_to change(User, :count)
-    end
+    click_button "Googleでログイン"
+
+    expect(page).to have_content "今どんなことが起きていますか？"
   end
 end
